Added a base file for the main stage code and added another layer of import redirection
