+++ pyproject.toml
@@ -41,16 +41,16 @@ MEDS_transform-stage = "MEDS_transforms.__main__:run_stage"
 MEDS_transform-pipeline = "MEDS_transforms.runner:main"
 
 [project.entry-points."MEDS_transforms.stages"]
-aggregate_code_metadata = "MEDS_transforms.stages.aggregate_code_metadata:main"
-fit_vocabulary_indices = "MEDS_transforms.stages.fit_vocabulary_indices:main"
-reshard_to_split = "MEDS_transforms.stages.reshard_to_split:main"
-filter_measurements = "MEDS_transforms.stages.filter_measurements:main"
-filter_subjects = "MEDS_transforms.stages.filter_subjects:main"
-reorder_measurements = "MEDS_transforms.stages.reorder_measurements:main"
-add_time_derived_measurements = "MEDS_transforms.stages.add_time_derived_measurements:main"
-extract_values = "MEDS_transforms.stages.extract_values:main"
-normalization = "MEDS_transforms.stages.normalization:main"
-occlude_outliers = "MEDS_transforms.stages.occlude_outliers:main"
+aggregate_code_metadata = "MEDS_transforms.stages:aggregate_code_metadata"
+fit_vocabulary_indices = "MEDS_transforms.stages:fit_vocabulary_indices"
+reshard_to_split = "MEDS_transforms.stages:reshard_to_split"
+filter_measurements = "MEDS_transforms.stages:filter_measurements"
+filter_subjects = "MEDS_transforms.stages:filter_subjects"
+reorder_measurements = "MEDS_transforms.stages:reorder_measurements"
+add_time_derived_measurements = "MEDS_transforms.stages:add_time_derived_measurements"
+extract_values = "MEDS_transforms.stages:extract_values"
+normalization = "MEDS_transforms.stages:normalization"
+occlude_outliers = "MEDS_transforms.stages:occlude_outliers"
 
 [project.urls]
 Homepage = "https://github.com/mmcdermott/MEDS_transforms"
